Recalculate boid neighborhood more infrequently
Calculating each boid’s neighborhood is computationally expensive and doesn’t need to be executed 60 times for second.  Factored this out so that it can be executed a couple of times per second instead, which drastically increases the number of boids that can appear on screen at any one time with no visual side effects.